third_party/blink/web_tests/platform/generic/http/tests/devtools/profiler/heap-snapshot-summary-expand-collapse-expected.txt

https://crbug.com/738932 Tests the snapshot view is not empty on repeatitive expand-collapse.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testShowAll

Profiler was disabled.
